New upload of my Web Development Class project.
Still learning Git Hub and I wanted to change the name of the folder in my projects library and wasn't successful at linking it to Git Hub so I uploaded it again with the proper name. I will get used to naming conventions and become consistent with time and practice.